task2/requirements.txt
# ================= RAG SYSTEM (STABLE) =================

# Database
pymongo==4.6.1

# Data Processing
pandas==2.1.4
openpyxl==3.1.2
numpy==1.26.3

# Vector Store
faiss-cpu==1.7.4

# Embeddings
sentence-transformers==2.2.2
huggingface_hub==0.16.4

# Transformers & Torch (CPU compatible)
transformers==4.30.2
torch==2.1.2

# LLM
groq==0.4.2

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
